feat: add Timestamp to messages
A Timestamp is added to each message generated by a Source. If a Source
generates several outputs, the same timestamp is added to all outputs.

As of this implementation, an Operator propagates the most recent
timestamp: if an Operator has several inputs, it will propagate with all its
outputs the most recent timestamp of all inputs.

A Timestamp is retrieved by an Operator via the `ReadyToken`. If an Operator
only processes `NotReadyToken` (if ever that is possible…), the Operator will
generate a Timestamp and attach it to all its outputs.